% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.rank.approx.R
\name{approx_rank_mutual}
\alias{approx_rank_mutual}
\title{Approximation of mutual rank probabilities}
\usage{
approx_rank_mutual(P)
}
\arguments{
\item{P}{a partial order as matrix object}
}
\value{
a matrix containing approximation of mutual rank probabilities
}
\description{
Mutual Rank approximation
}
\details{
TODO
}
\examples{
###TODO
}
\seealso{
\link{rank_analysis}
}
