% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.ranks.R
\name{approx_rank_relative}
\alias{approx_rank_relative}
\title{Approximation of relative rank probabilities}
\usage{
approx_rank_relative(P, iterative = T, num.iter = 10)
}
\arguments{
\item{P}{a partial order as matrix object}

\item{iterative}{boolean. TRUE(default) if iterative approximation should be used. FALSE if not.}

\item{num.iter}{number of iterations to be used. defaults to 10 (see Details)}
}
\value{
a matrix containing approximation of mutual rank probabilities.
\code{relative.rank[i,j]} is the probability that i is ranked lower than j
}
\description{
approximate relative rank probabilites
}
\details{
The iterative approach generally gives better approximations than the non iterative, yet only slightly.
More than 10 iterations do not improve the accuracy significantly.
}
\examples{
P=matrix(c(0,0,1,1,1,0,0,0,1,0,0,0,0,0,1,rep(0,10)),5,5,byrow=TRUE)
P
approx_rank_relative(P,iterative=FALSE) 
approx_rank_relative(P,iterative=TRUE)
}
\seealso{
\link{rank_analysis}
}
