% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.matrix.netrankr_full}
\alias{as.matrix.netrankr_full}
\title{Extract probabilities from netrankr_full object}
\usage{
\method{as.matrix}{netrankr_full}(x, type = "rank", ...)
}
\arguments{
\item{x}{A netrankr_full object}

\item{type}{which probabilities to return. "rank" for rank probabilities, "relative" for relative rank probabilities and "expected" for expected rank probabilities and their variants}

\item{...}{additional parameters for as.matrix}
}
\description{
extract probabilities as matrices from the result of an object obtained from \link{exact_rank_prob}
}
\author{
David Schoch
}
