% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley.terry.R
\name{bradley_terry}
\alias{bradley_terry}
\title{Bradley Terry Model}
\usage{
bradley_terry(P, sparse.correct = 0, max.iter = 100, tol = 10^-8,
  print.level = 0)
}
\arguments{
\item{P}{a matrix with partial order information}

\item{sparse.correct}{eps to correct for weak connectivity}

\item{max.iter}{integer, maximum number of MLE iterations}

\item{tol}{double, convergence criterion}

\item{print.level}{binary, should diagnostics be printed or not}
}
\value{
data frame
}
\description{
Calculate merrits from dominance relations via BM
}
\details{
TODO
}
\examples{
###TODO
}
