% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.preservation.R
\name{check_preservation}
\alias{check_preservation}
\title{Checks if a partial ranking is preserved}
\usage{
check_preservation(P, scores)
}
\arguments{
\item{P}{a partial ranking as a matrix object}

\item{scores}{scores for each element}
}
\value{
a boolean
}
\description{
Checks if a partial ranking is preserved in the ranking induced by scores
}
\details{
In order for a score vector to preserve a partial ranking, the following
condition must be fulfilled:
\code{P[u,v]==1 & scores[i]<=scores[j]}
}
\examples{
require(igraph)
g=graph.star(5,"undirected")
neighborhood_inclusion(g)
}
