% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.rank.R
\name{color_rank}
\alias{color_rank}
\title{Color Paths in Ideal Lattice}
\usage{
color_rank(g, path.label, col = "red")
}
\arguments{
\item{g}{igraph object. Lattice of ideals}

\item{path.label}{vector of vertex ids. Ranking to be colorized}

\item{col}{color to be used}
}
\value{
igraph object with edge color attribute
}
\description{
Simple Function to color rankings in the lattice of ideal
}
\examples{
###TODO
}
