% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparable.pairs.R
\name{comparable_pairs}
\alias{comparable_pairs}
\title{Comparable pairs of a partial order}
\usage{
comparable_pairs(P)
}
\arguments{
\item{P}{matrix containing comparability relation}
}
\value{
fraction of comparable pairs in P
}
\description{
Calculates the fraction of comparable pairs in a partial order. This fraction is identical to
the density of the induced undirected graph of P.
}
\examples{
require(igraph)
g=erdos.renyi.game(100,0.1)
P=neighborhood_inclusion(g)
comparable_pairs(P)
}
