% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.ranks.R
\name{compare_ranks}
\alias{compare_ranks}
\title{Count occurences of pairs in rankings}
\usage{
compare_ranks(x, y)
}
\arguments{
\item{x}{a ranking vector}

\item{y}{a ranking vector with the same length as \emph{x}}
}
\value{
a list containing
\item{c}{number of concordant pairs: \code{x[i]}>\code{x[j]} and \code{y[i]}>\code{y[j]}}
\item{d}{number of discordant pairs: \code{x[i]}>\code{x[j]} and \code{y[i]}<\code{y[j]}}
\item{t}{number of tied pairs:  \code{x[i]}=\code{x[j]} and \code{y[i]}=\code{y[j]}}
\item{u}{number of left and right ties: \code{x[i]}=\code{x[j]} and \code{y[i]}>\code{y[j]}}
}
\description{
Counts the number of Concordant, Discordant, Tied and Unknown between two rankings
}
\details{
TODO
}
\examples{
require(igraph)
tg=threshold_graph(100,0.2)
compare_ranks(degree(tg),closeness(tg)) #only concordant pairs
compare_ranks(degree(tg),betweenness(tg)) #no discordant pairs
}
