% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbces11}
\alias{dbces11}
\title{dbces11 graph}
\format{
igraph object
}
\usage{
dbces11
}
\description{
Smallest graph (11 nodes and 17 edges) where the centers according to (d)egree, (b)etweenness, (c)loseness, (e)igenvector centrality, and (s)ubgraph centrality are all different.
}
\keyword{datasets}
