% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.transform.R
\name{dependency_transform}
\alias{dependency_transform}
\title{Dyadic Dependencies between nodes}
\usage{
dependency_transform(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
a matrix containing dyadic dependencies
}
\description{
Calculates dyadic dependencies
\deqn{
\delta(s\vert u)=\sum_t \frac{\sigma(s,t\vert u)}{\sigma(s,t)}
}
between pairs of nodes. The values are commonly used for betweenness centrality.
}
\details{
Betweenness centrality is equal to column sums!
}
\examples{
###TODO
}
\seealso{
\link{positional_dominance}
}
