% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.index.R
\name{distance_index}
\alias{distance_index}
\title{Distance-based centrality indices}
\usage{
distance_index(g, type = "ros", alpha = 1)
}
\arguments{
\item{g}{An igraph object}

\item{type}{character specifying the used index. See details for possibilities}

\item{alpha}{numeric value for parametrized indices. See details}
}
\value{
Numeric vector with scores of the specified index
}
\description{
Implements a variety of centrality indices that are based on geodesic distances.
}
\details{
The \emph{type} parameter can be set to
\describe{
\item{ros}{"reciprocal of sum". \deqn{c(u)=\frac{1}{\sum_{t \in V} dist(i,t)}}}
\item{sor}{"sum of reciprocals". \deqn{c(u)=\sum_{t \in V} \frac{1}{dist(i,t)}}}
\item{dpow}{"distance to the power of". \deqn{c(u)=\sum_{t \in V} dist(i,t)^{-\alpha} \quad \alpha\geq 0}}
\item{powd}{"to the power of distance". \deqn{c(u)=\sum_{t \in V} \alpha^{-dist(i,t)} \quad \alpha \in (0,1)}}
\item{int}{"integration". \deqn{c(u)=\sum_{t \in V} 1-\frac{dist(i,t)-1}{diameter(G)}}}
\item{pow2}{"distance to the power of 2". \deqn{c(u)=\sum_{t \in V} 2^{-dist(i,t)}}}
}
}
\examples{
require(igraph)
g=sample_gnp(100,0.1)
d_ros <- distance_index(g,type="ros")
d_sor <- distance_index(g,type="sor")
compare_ranks(d_ros,d_sor) #compare rankings
}
