% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.transform.R
\name{distance_transform}
\alias{distance_transform}
\title{Distances between nodes}
\usage{
distance_transform(g, mode = "all")
}
\arguments{
\item{g}{igraph object}

\item{mode}{Character constant, gives whether the distances to or from the given vertices
should be calculated for directed graphs.
If out then the shortest paths from the vertex,
if in then to it will be considered.
If all, the default, then the corresponding undirected graph will be used, ie. not directed paths are searched.}
}
\value{
a matrix containing distances
}
\description{
Calculates distances between pairs of nodes. The values are commonly used for closeness centrality.
}
\details{
#TODO
}
\examples{
###TODO
}
\seealso{
\link{positional_dominance}
}
