% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.graph.R
\name{dominance_graph}
\alias{dominance_graph}
\title{Neighborhood Inclusion Graph}
\usage{
dominance_graph(g)
}
\arguments{
\item{g}{igraph object}
}
\value{
directed igraph object with dominance relations
}
\description{
Displays the neighborhood inclusion preorder as a directed graph
}
\examples{
require(igraph)
g <- erdos.renyi.game(100,0.1)
d <- dominance_graph(g)
plot(d)
}
\author{
David Schoch
}
