% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florentine_m.R
\docType{data}
\name{florentine_m}
\alias{florentine_m}
\title{Florentine Family marriage network}
\format{An igraph object containing marriage links of florentine families}
\usage{
florentine_m
}
\description{
Florentine Family marriage network
}
\keyword{datasets}
