% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.ranks.R
\name{freeman_hierarchy}
\alias{freeman_hierarchy}
\title{Freemans Hierarchy measure}
\usage{
freeman_hierarchy(P)
}
\arguments{
\item{P}{a partial order as matrix object}
}
\value{
a matrix with four columns containing Freemans hierarchy.
}
\description{
A Freeman's hierarchy is based on the singular value decomposition of the skew-symmetric matrix
\deqn{Z=\frac{P^T-P}{2}}{(P^T-P)/2}
}
\details{
TODO
}
\examples{
###TODO
}
\seealso{
\link{rank_analysis},\link{approx_rank_expected}
}
