% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.builder.R
\name{index_builder}
\alias{index_builder}
\title{Centrality Index Builder}
\usage{
index_builder()
}
\value{
code to calculate the specified index.
}
\description{
This shiny gadget can be used to build centrality indices based on specific indirect relations,
transformations and aggregation functions. use the dropdown menus to select
components that make up the index. Depending on your choices, some options
are not available at later stages. At the end, code is being inserted into
the current script to use the index
}
