% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.rank.R
\name{mcmc_rank_prob}
\alias{mcmc_rank_prob}
\title{Estimate Rankings with MCMC}
\usage{
mcmc_rank_prob(P, rp = 10000)
}
\arguments{
\item{P}{a matrix representing a partial ranking.}

\item{rp}{integer indicating the number of samples to be drawn.}
}
\value{
data frame containing approximated expected ranks and relative rank probabilities.

#' @seealso \link{exact_rank_prob}, \link{approx_rank_relative}, \link{approx_rank_expected}
}
\description{
Performs a probabilistic rank analysis based on an almost uniform
sample of possible rankings that preserve the partial ranking, instead of taking all possibilities
as in \link{exact_rank_prob}.
}
\details{
As a rule of thumb, the number of samples should at least be cubic in
the number of nodes of the network, but the higher the better.
}
\examples{
###TODO
}
\author{
David Schoch
}
