% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.netrankr_mcmc}
\alias{plot.netrankr_mcmc}
\title{Plot netrankr_mcmc object}
\usage{
\method{plot}{netrankr_mcmc}(x, icols = hcl.colors(12, "viridis"), bcol = "grey66", ...)
}
\arguments{
\item{x}{A netrankr_mcmc object}

\item{icols}{a list of colors such as that generated by \link{hcl.colors}, \link{gray.colors} or similar functions for the heatmaps}

\item{bcol}{color used for the barcharts}

\item{...}{additional plot parameters}
}
\description{
Plots the result of an object obtained from \link{mcmc_rank_prob}
}
\author{
David Schoch
}
