% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rank.intervals.R
\name{plot_rank_intervals}
\alias{plot_rank_intervals}
\title{Plot Rank Intervals}
\usage{
plot_rank_intervals(P, names, cent.df)
}
\arguments{
\item{P}{A partial ranking as matrix for which rank intervals should be determined.}

\item{names}{string. potential names of nodes used in the plot (optional)}

\item{cent.df}{a data frame containing centrality scores of indices (optional). See Details.}
}
\value{
a ggplot object.
}
\description{
Derive indirect relations like distances for a given network.
}
\details{
The
}
\examples{
require(igraph)
require(ggplot2)
g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
P <- neighborhood_inclusion(g)
plot_rank_intervals(P)

#adding index based rankings
cent_scores <- data.frame(
  degree=degree(g),
  betweenness=round(betweenness(g),4),
  closeness=round(closeness(g),4),
  eigenvector=round(eigen_centrality(g)$vector,4))
plot_rank_intervals(P,cent.df=cent_scores)
}
\author{
David Schoch
}
