% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional.dominance.R
\name{positional_dominance}
\alias{positional_dominance}
\title{Generalized Dominance in Graphs}
\usage{
positional_dominance(A, map = F, benefit = T)
}
\arguments{
\item{A}{matrix containing attributes}

\item{map}{boolean if rows can be sorted or not(default)}

\item{benefit}{boolean if higher values(default) or lower values are better}
}
\value{
dominance relations
}
\description{
generealized dominance relations. More to come
}
\examples{
###TODO
}
