% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.netrankr_full}
\alias{print.netrankr_full}
\title{Print netrankr_full object to terminal}
\usage{
\method{print}{netrankr_full}(x, ...)
}
\arguments{
\item{x}{A netrank object}

\item{...}{additional arguments to print}
}
\description{
Prints the result of an object obtained from \link{exact_rank_prob} to terminal
}
\author{
David Schoch
}
