% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.netrankr_interval}
\alias{print.netrankr_interval}
\title{Print netrankr_interval object to terminal}
\usage{
\method{print}{netrankr_interval}(x, ...)
}
\arguments{
\item{x}{A netrankr_interval object}

\item{...}{additional arguments to print}
}
\description{
Prints the result of an object obtained from \link{rank_intervals} to terminal
}
\author{
David Schoch
}
