% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.netrankr_mcmc}
\alias{print.netrankr_mcmc}
\title{Print netrankr_mcmc object to terminal}
\usage{
\method{print}{netrankr_mcmc}(x, ...)
}
\arguments{
\item{x}{A netrank object}

\item{...}{additional arguments to print}
}
\description{
Prints the result of an object obtained from \link{mcmc_rank_prob} to terminal
}
\author{
David Schoch
}
