% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.analysis.R
\name{rank_analysis}
\alias{rank_analysis}
\title{Rank Analysis of networks}
\usage{
rank_analysis(P, names = "", only.results = T, verbose = F)
}
\arguments{
\item{P}{matrix representing a partial ranking}

\item{names}{optional argument for names if P does not have row/column names}

\item{only.results}{logical. wether only results (default) or additionally the ideal tree and lattice should be returned}

\item{verbose}{logical. should diagnostics be printed. Defaults to F}
}
\value{
\item{n.ext}{Number of possible centrality rankings}
\item{topo.order}{Random Ranking used to build the lattice of ideals}
\item{mse}{Array indicating equivalent nodes}
\item{rank.prob}{Matrix containing rank probabilities: \code{rank.prob[i,k]} is the probability that i has rank k}
\item{mutual.rank}{Matrix containing relative rank probabilities: \code{mutual.rank[i,j]} is the probability that i is ranked lower than j}
\item{expected.rank}{Expected ranks of nodes in any centrality ranking}
\item{rank.spread}{Variance of the ranking probabilities}
\item{lattice}{igraph object. The lattice of ideals (if only.results=F)}
\item{tree}{igraph object. The tree of ideals (if only.results=F)}
}
\description{
Performes a complete Rank analysis of a network.
Calculates Expected Rankings, (Relative) Rank Probabilities and number of possible rankings.
Works best with small networks.
}
\details{
TODO
}
\examples{
###TODO
}
\references{
\insertRef{ddd-elirp-06}{netrankr}
}
\seealso{
\link{approx_rank_mutual}, \link{approx_rank_expected}
}
