% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.plots.R
\name{rank_plots}
\alias{rank_plots}
\title{Visualize expected ranks and rank probabilities}
\usage{
rank_plots(df, method = "")
}
\arguments{
\item{df}{a dataframe as returned from rank.analysis}

\item{method}{string, see details}
}
\value{
ggplot object
}
\description{
Uses the output of
Works best with small networks
}
\details{
TODO
}
\examples{
###TODO
}
