% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.netrankr_full}
\alias{summary.netrankr_full}
\title{Summary of a netrankr_full object}
\usage{
\method{summary}{netrankr_full}(object, ...)
}
\arguments{
\item{object}{A netrankr_full object}

\item{...}{additional arguments to summary}
}
\description{
Summarizes the result of an object obtained from \link{exact_rank_prob} to terminal
}
\author{
David Schoch
}
