% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.relations.R
\name{transform_relations}
\alias{transform_relations}
\alias{walks_limit_prop}
\alias{walks_even_exp}
\alias{walks_odd_exp}
\alias{walks_exp}
\alias{walks_attenuated}
\alias{walks_uptok}
\alias{dist_2pow}
\alias{dist_inv}
\alias{dist_dpow}
\alias{dist_powd}
\title{transformation of indirect relations}
\usage{
walks_limit_prop(x)

walks_even_exp(x, alpha = 1)

walks_odd_exp(x, alpha = 1)

walks_exp(x, alpha = 1)

walks_attenuated(x, alpha = 1/max(x) * 0.99)

walks_uptok(x, alpha = 1, k = 3)

dist_2pow(x)

dist_inv(x)

dist_dpow(x, alpha = 1)

dist_powd(x, alpha = 0.5)
}
\arguments{
\item{x}{matrix of transformed relations.}

\item{alpha}{weighting factor.}

\item{k}{for walk counts up to a certain length.}
}
\value{
transformed relations as matrix
}
\description{
Mostly wrapper functions to transform indirect relations.

This functions can be used in conjunction with \link{indirect_relations}.
}
\author{
David Schoch
}
