% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.ranks.R
\name{approx_rank_expected}
\alias{approx_rank_expected}
\title{Approximation of expected rankings}
\usage{
approx_rank_expected(P, method = "lpom")
}
\arguments{
\item{P}{a partial order as matrix object}

\item{method}{string indicating method to be used. see details}
}
\value{
a vector containing approximations of expected ranks
}
\description{
Implements a variety of functions to approximate expected ranks in large partially ordered sets
}
\details{
The \emph{method} parameter can be set to
\describe{
\item{lpom}{ simplest, most accurate for tiny networks}
\item{glpom}{ most accurate for tiny networks}
\item{loof1}{ more accurate for large networks}
\item{loof2}{ even more accurate for large networks}
}
}
\examples{
###TODO
}
\seealso{
\link{rank_analysis},\link{bradley_terry}
}
