% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.ranks.R
\name{approx_rank_relative}
\alias{approx_rank_relative}
\title{Approximation of relative rank probabilities}
\usage{
approx_rank_relative(P, iterative = T, num.iter = 10)
}
\arguments{
\item{P}{a partial order as matrix object}

\item{iterative}{boolean. TRUE(default) if iterative approximation should be used. FALSE if not.}

\item{num.iter}{number of iterations to be used. defaults to 10 (see Details)}
}
\value{
a matrix containing approximation of mutual rank probabilities.
\code{mutual.rank[i,j]} is the probability that i is ranked lower than j
}
\description{
Relative Rank approximation
}
\details{
The iterative approach generally gives better approximations than the non iterative, yet only slightly.
more than 10 iterations do not seem to significantly improve the accuracy
}
\examples{
###TODO
}
\seealso{
\link{rank_analysis}
}
