% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.ranks.R
\name{approx_rank_relative}
\alias{approx_rank_relative}
\title{Approximation of relative rank probabilities}
\usage{
approx_rank_relative(P, iterative = TRUE, num.iter = 10)
}
\arguments{
\item{P}{a partial order as matrix object.}

\item{iterative}{boolean. TRUE (default) if iterative approximation should be used. FALSE if not.}

\item{num.iter}{number of iterations to be used. defaults to 10 (see Details).}
}
\value{
a matrix containing approximation of mutual rank probabilities.
\code{relative.rank[i,j]} is the probability that i is ranked lower than j
}
\description{
Approximate relative rank probabilities \eqn{Prob(u<v)}. In a network context, \eqn{Prob(u<v)}
gives the probability that u is less central than v, given the partial ranking P.
}
\details{
The iterative approach generally gives better approximations than the non iterative, if only slightly.
The default number of iterations is based on the observation, that the approximation does not improve
significantly beyond this value. This observation, however, is based on very small networks such that
increasing it for large network may yield better results.
}
\examples{
P=matrix(c(0,0,1,1,1,0,0,0,1,0,0,0,0,0,1,rep(0,10)),5,5,byrow=TRUE)
P
approx_rank_relative(P,iterative=FALSE) 
approx_rank_relative(P,iterative=TRUE)
}
\seealso{
\link{exact_rank_prob} for exact computations
}
