% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.preservation.R
\name{check_preservation}
\alias{check_preservation}
\title{Checks if a partial ranking is preserved}
\usage{
check_preservation(P, scores)
}
\arguments{
\item{P}{a partial ranking as a matrix object.}

\item{scores}{scores for each element.}
}
\value{
a boolean that is TRUE if \code{scores} preserves the relations in \code{P}.
}
\description{
Checks if a partial ranking is preserved in the ranking induced by scores.
}
\details{
In order for a score vector to preserve a partial ranking, the following
condition must be fulfilled:
\code{P[u,v]==1 & scores[i]<=scores[j]}.
}
\examples{

require(igraph)
### standard measures of centrality preserve the neighborhood inclusion preorder
g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
P<-neighborhood_inclusion(g)

check_preservation(P,degree(g))
check_preservation(P,betweenness(g))
check_preservation(P,closeness(g))
}
