% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.ranks.R
\name{compare_ranks}
\alias{compare_ranks}
\title{Count occurences of pairs in rankings}
\usage{
compare_ranks(x, y)
}
\arguments{
\item{x}{a ranking vector.}

\item{y}{a ranking vector with the same length as \code{x}.}
}
\value{
a list containing
\item{concordant}{number of concordant pairs: \code{x[i]}>\code{x[j]} and \code{y[i]}>\code{y[j]}}
\item{discordant}{number of discordant pairs: \code{x[i]}>\code{x[j]} and \code{y[i]}<\code{y[j]}}
\item{ties}{number of tied pairs:  \code{x[i]}=\code{x[j]} and \code{y[i]}=\code{y[j]}}
\item{left}{number of left: \code{x[i]}=\code{x[j]} and \code{y[i]}!=\code{y[j]}}
\item{right}{number of right: \code{x[i]}!=\code{x[j]} and \code{y[i]}=\code{y[j]}}
}
\description{
Counts the number of Concordant, Discordant and (Left/Right) Ties between two rankings.
}
\details{
Explicitly calculating the number of occuring cases is more robust than using correlation
indices as given in the \code{cor} function. Especially left and right ties can significantly alter an
intuitively high correlation (See example).
}
\examples{
require(igraph)
tg <- threshold_graph(100,0.2)
compare_ranks(degree(tg),closeness(tg)) #only concordant pairs
compare_ranks(degree(tg),betweenness(tg)) #no discordant pairs
## Rank Correlation
cor(degree(tg),closeness(tg)) #1
cor(degree(tg),betweenness(tg)) #not 1, although no discordant pairs
}
