% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.ranks.R
\name{compare_ranks}
\alias{compare_ranks}
\title{Count occurences of pairs in rankings}
\usage{
compare_ranks(x, y)
}
\arguments{
\item{x}{a ranking vector}

\item{y}{a ranking vector with the same length as \emph{x}}
}
\value{
a list containing
\item{c}{number of concordant pairs}
\item{d}{number of discordant pairs}
\item{t}{number of tied pairs}
\item{u}{number of left and right ties}
}
\description{
Counts the number of Concordant, Discordant, Tied and Unknown between two rankings
}
\details{
This function allows
}
\examples{
###TODO
}
