% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.graph.R
\name{dominance_graph}
\alias{dominance_graph}
\title{Directed Comparability Graph}
\usage{
dominance_graph(P)
}
\arguments{
\item{P}{Partial ranking as matrix object}
}
\value{
igraph object.
}
\description{
Turns a partial ranking into a directed graph.
}
\details{
An edge (u,v) is present if u is dominated by v in P.
}
\examples{
require(igraph)
g <- threshold_graph(20,0.1)
P <- neighborhood_inclusion(g)
d <- dominance_graph(P)
plot(d)
# to reduce overplotting use transitive reduction
P <- transitive_reduction(P)
d <- dominance_graph(P)
plot(d)
}
\author{
David Schoch
}
