% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florentine_m.R
\docType{data}
\name{florentine_m}
\alias{florentine_m}
\title{Florentine family marriage network}
\format{
An igraph object containing marriage links of florentine families
}
\usage{
florentine_m
}
\description{
Florentine family marriage network
}
\references{
Padgett, J.F. and Ansell, C.K., 1993.
Robust Action and the Rise of the Medici, 1400-1434. \emph{American Journal of Sociology}, \strong{98}(6), 1259-1319.
}
\keyword{datasets}
