% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{All rankings of a partial ranking}
\usage{
get_rankings(lattice, ideals, topo.order, linext, mse, force = F)
}
\arguments{
\item{lattice}{adjacency list of predecessors in lattice of ideal. This list is returned if \link{rank_analysis}
is run with \code{only.results=FALSE}}

\item{ideals}{list of ideals. This list is returned if \link{rank_analysis}
is run with \code{only.results=FALSE}}

\item{topo.order}{topological order used in \link{rank_analysis}}

\item{linext}{number of possible rankings. returned by rank_analysis}

\item{mse}{equivalence classes of rankings. returned by \link{rank_analysis}}

\item{force}{boolean. stops function if the number of rankings is too large.
Only change to TRUE if you know what you are doing}
}
\value{
a matrix containing all ranking positions of nodes
}
\description{
Returns all possible rankings of a partial ranking using the lattice of ideals.
}
\examples{
P <- matrix(c(0,0,1,1,1,0,0,0,1,0,0,0,0,0,1,rep(0,10)),5,5,byrow=TRUE)
P
res <- rank_analysis(P,only.results=FALSE)
get_rankings(res$lattice,res$ideals,res$topo.order,res$lin.ext,res$mse)

}
\author{
David Schoch
}
