% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.builder.R
\name{index_builder}
\alias{index_builder}
\title{Centrality Index Builder}
\usage{
index_builder()
}
\value{
code to calculate the specified index.
}
\description{
This gadget can be used to build centrality indices based on specific indirect relations,
transformations and aggregation functions.
}
