% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indirect.relations.R
\name{indirect_relations}
\alias{indirect_relations}
\title{Indirect relations of a network}
\usage{
indirect_relations(g, relation = "distances", walk_length = 10,
  gamma = factorial(1:walk_length))
}
\arguments{
\item{g}{igraph object. The network for which relations should be derived}

\item{relation}{string. giving the relation to be calculated. See details for options.}

\item{walk_length}{integer. maximum length of walks to be considered if relation="walks"}

\item{gamma}{vector. weighting of walks if relation="walks".}
}
\value{
a matrix containing the derived relations
}
\description{
Derive indirect relations, for instance distances, for a given network.
}
\details{
the following relations
}
\examples{

require(igraph)
### standard measures of centrality preserve the neighborhood inclusion preorder
g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
dist <- indirect_relations(g,"distances") #same as distances(g,mode="all")
dep  <- indirect_relations(g,"dependencies")

#rowSums of dep equals 2*betweenness(g)
rowSums(dep)-2*betweenness(g)

#indirect realtion used by subgraph_centrality
walk  <- indirect_relations(g,"walks",gamma=factorial(1:10))
}
