% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorization.gap.R
\name{majorization_gap}
\alias{majorization_gap}
\title{(normalized) Majorization Gap of a Graph}
\usage{
majorization_gap(g, norm = TRUE)
}
\arguments{
\item{g}{An igraph object}

\item{norm}{True(Default) if the normalized majorization gap should be returned.}
}
\value{
majorization gap of an undirected graph.
}
\description{
Calculates the (normalized) majorization gap of a graph. The majorization gap indicates how far the
degree sequence of a graph is from a threshold sequence. The distance is measured by the number of \emph{reverse unit
transformations} necessaray to turn the degree sequence into a threshold sequence.
}
\examples{
require(igraph)
g=graph.star(5,"undirected")
majorization_gap(g) #0 since star graphs are threshold graphs

g=sample_gnp(100,0.15)
majorization_gap(g,norm=TRUE) #fraction of reverse unit transformation
majorization_gap(g,norm=FALSE) #number of reverse unit transformation
}
