% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.inclusion.R
\name{neighborhood_inclusion}
\alias{neighborhood_inclusion}
\title{Neighborhood-Inclusion Preorder}
\usage{
neighborhood_inclusion(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
the neighborhood-inclusion preorder of an undirected graph g
}
\description{
Calculates the neighborhood-inclusion preorder of an undirected graph
}
\examples{
require(igraph)
g=graph.star(5,"undirected")
neighborhood_inclusion(g)
}
