% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.inclusion.R
\name{neighborhood_inclusion}
\alias{neighborhood_inclusion}
\title{Neighborhood-Inclusion Preorder}
\usage{
neighborhood_inclusion(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
the neighborhood-inclusion preorder of an undirected graph g as matrix object
}
\description{
Calculates the neighborhood-inclusion preorder of an undirected graph.
}
\details{
neighborhood-inclusion is formally defined as
\deqn{N(u)\subseteq N[v]}
where \eqn{N(u)} is the neighborhood of \eqn{u} and \eqn{N[v]=N(v)\cup \lbrace v\rbrace} is the closed neighborhood of \eqn{v}.
\eqn{N(u) \subseteq N[v]} implies that \eqn{c(u) \leq c(v)},
where \eqn{c} is an arbitrary centrality index. Neighborhood-inclusion is thus preserved in any centrality ranking.
}
\examples{
require(igraph)
###the neighborhood inclusion preorder of a star graph is complete
g=graph.star(5,"undirected")
P <- neighborhood_inclusion(g)
comparable_pairs(P)

###standard centrality indices preserve the partial ranking of neighborhood inclusion
g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
P<-neighborhood_inclusion(g)
is_preserved(P,degree(g))
is_preserved(P,closeness(g))
is_preserved(P,betweenness(g))
}
\references{
Brandes, U. Heine, M., Müller, J. and Ortmann, M., 2017.
Positional Dominance: Concepts and Algorithms. \emph{Conference on Algorithms and Discrete Applied Mathematics}, 60-71.

Schoch, D. and Brandes, U., 2016. Re-conceptualizing centrality in social networks.
\emph{European Journal of Applied Mathematics} 27(6), 971-985.
}
\seealso{
\link{positional_dominance}, \link{is_preserved}
}
\author{
David Schoch
}
