% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrankr.R
\docType{package}
\name{netrankr}
\alias{netrankr}
\alias{netrankr-package}
\title{netrankr: Network centrality (almost) without indices.}
\description{
netrankr provides several functions for centrality related analyses of networks.
The main focus lies on methods that do not rely on indices like degree,
betweenness or closeness.
}
\details{
Some notable features of the package are:

\itemize{
\item Working with the neighborhood inclusion preorder. This forms the bases
for any centrality analysis on undirected and unweighted graphs.
More details can be found in the dedicated vignette:
\code{vignette("neighborhood_inclusion",package="netrankr")}
\item Constructing graphs with a unique centrality ranking.
This class of graphs, known as threshold graphs, can be used to benchmark
centrality indices, since they only allow for one ranking of the nodes.
For more details consult the vignette: \code{vignette("threshold_graph",package="netrankr")}
\item Probabilistic centrality. Why applying a handful of indices and choosing
the one that fits best, when it is possible to analyze \strong{all} centrality rankings at once?
The package includes several function to
calculate rank probabilities of nodes in a network. These include expected ranks and
relative rank probabilities (how likely is it that a node is more central than another?)
Consult \code{vignette("probabilistic_cent",package="netrankr")} for more info.
}

The package provides many additional vignettes that explain the functionality
of netrankr and its conceptual ideas. See \code{browseVignettes(package = "netrankr")}
}
