% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional.dominance.R
\name{positional_dominance}
\alias{positional_dominance}
\title{Generalized Dominance in Graphs}
\usage{
positional_dominance(A, map = FALSE, benefit = TRUE)
}
\arguments{
\item{A}{matrix containing attributes or relations}

\item{map}{boolean if rows can be sorted or not (default)}

\item{benefit}{boolean if higher values(default) or lower values are better}
}
\value{
dominance relations as matrix object.
}
\description{
generalized dominance relations.
}
\details{
Positional dominance is a generalization of neighborhood-inclusion for
arbitrary network data. In the default case, it checks for all pairs \eqn{i,j} if
\eqn{A_{it} \geq A_{jt}} holds for all \eqn{t}. If \code{map=TRUE},
the rows of \eqn{A} are sorted decreasingly (\code{benefit=TRUE}) or increasingly
(\code{benefit=FALSE}) and then the dominance condition is checked.
}
\examples{
###
require(igraph)

g <- graph.empty(n=11,directed = FALSE)
g <- add_edges(g,c(1,11,2,4,3,5,3,11,4,8,5,9,5,11,6,7,6,8,
                   6,10,6,11,7,9,7,10,7,11,8,9,8,10,9,10))
P<-neighborhood_inclusion(g)
comparable_pairs(P)

dist <- distances(g)
D <- positional_dominance(dist,map=FALSE,benefit=FALSE) 
comparable_pairs(D) #same as P

D <- positional_dominance(dist,map=TRUE,benefit=FALSE) 
comparable_pairs(D) #more comparables than P

### all distance based indices preserve positional dominance
check_preservation(D,distance_index(g,type="sor"))
check_preservation(D,distance_index(g,type="ros"))
check_preservation(D,distance_index(g,type="pow2"))
check_preservation(D,distance_index(g,type="int"))
}
\seealso{
\link{neighborhood_inclusion}, \link{check_preservation}
}
\author{
David Schoch
#' @references Brandes, U. Heine, M., Müller, J. and Ortmann, M., 2017.
Positional Dominance: Concepts and Algorithms. \emph{Conference on Algorithms and Discrete Applied Mathematics}, 60-71.

Schoch, D. and Brandes, U., 2016. Re-conceptualizing centrality in social networks.
\emph{European Journal of Applied Mathematics} 27(6), 971-985.
}
