% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.analysis.R
\name{rank_analysis}
\alias{rank_analysis}
\title{Rank Analysis of networks}
\usage{
rank_analysis(P, method = "bfs", names = "", print.level = 0)
}
\arguments{
\item{P}{a partial order as matrix object}

\item{method}{string indicating method to get all possible rankings. Default should not be changed}

\item{names}{optional argument for names if P does not have row/column names}

\item{print.level}{should diagnostics be printed (1) or not (0). Defaults to 0}
}
\value{
a list containing
\item{n.ext}{Number of Linear Extensions of P}
\item{mse}{Array indicating equivalent nodes}
\item{rank.prob}{Matrix containing rank probabilities \code{rank.prob[i,k]} is the probability that i has rank k}
\item{mutual.rank}{Matrix containing mutual rank probabilities. \code{mutual.rank[i,j]} is the probability that i is ranked higher than j}
\item{expected.rank}{Expected Ranking of nodes in any centrality ranking}
\item{rank.spread}{???}
\item{g.lattice}{igraph object. The lattice of ideals}
}
\description{
Performes a complete Rank analysis of a network.
Calculates Expected Rankings, (Mutual) Rank Probabilities and number of possible rankings.
Works best with small networks.
}
\details{
TODO
}
\examples{
###TODO
}
\references{
\insertRef{ddd-elirp-06}{netrankr}
}
