% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.interval.R
\name{rank_intervals}
\alias{rank_intervals}
\title{Rank interval of nodes}
\usage{
rank_intervals(P)
}
\arguments{
\item{P}{a partial ranking as a matrix}
}
\value{
a data frame with the minimal, maximal rank of each node together with the mid point of the two extrema.
}
\description{
Calculate the maximal and minmal rank possible for each node using the partial ranking P.
}
\details{
Note that the returned \code{mid_point} is not the same as the expected rank, for instance computed with \link{rank_analysis}.
It is simply the mid point between \code{min_rank} and \code{max_rank}.
}
\examples{
P <- matrix(c(0,0,1,1,1,0,0,0,1,0,0,0,0,0,1,rep(0,10)),5,5,byrow=TRUE)
rank_intervals(P)
}
\seealso{
\link{rank_analysis}
}
\author{
David Schoch
}
