% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.gap.R
\name{spectral_gap}
\alias{spectral_gap}
\title{Spectral Gap of a Graph}
\usage{
spectral_gap(g, method = "fraction")
}
\arguments{
\item{g}{igraph object}

\item{method}{a string, either "frac" or "abs"}
}
\value{
numeric value
}
\description{
The spectral gap of a graph is the absolute difference between the biggest and second biggest eigenvalue
of the adjacency matrix. To compare spectral gaps across networks, the fraction can be used
}
\examples{
#The spectral gap of a threshold graph is usually quite big
g=threshold_graph(10,0.3)
spectral_gap(g,method="fraction")
}
