% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.graph.R
\name{threshold_graph}
\alias{threshold_graph}
\title{Constructing Threshold Graphs}
\usage{
threshold_graph(n, p)
}
\arguments{
\item{n}{number of nodes}

\item{p}{fraction of dominating vertices. Equates approximately to the density of the graph}
}
\value{
a threshold graph as igraph object
}
\description{
A threshold graph is a graph where all centrality indices give the same ranking.
}
\examples{
g <- threshold_graph(10,0.3)
plot(g)

###Compare induced rankings
require(igraph)
degree(g)
closeness(g)
cor(degree(g),closeness(g),method="kendall")
}
\references{
\insertRef{mp-tgrt-95}{netrankr}
}
