% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitive.reduction.R
\name{transitive_reduction}
\alias{transitive_reduction}
\title{Transitive Reduction of Dominance Relations}
\usage{
transitive_reduction(P)
}
\arguments{
\item{P}{matrix containing relations}
}
\value{
transitive reduction of P
}
\description{
Calculates the transitive reduction of a partial ranking.
}
\examples{
require(igraph)
g <- threshold_graph(100,0.1)
P <- neighborhood_inclusion(g)
sum(P)
T <- transitive_reduction(P)
sum(T)
}
\author{
David Schoch
}
